data/scrap_server/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
cryptography

requests
selenium
webdriver-manager
pandas
bs4
tqdm
pymysql
sqlalchemy
boto3
scikit-learn
matplotlib